+++ src/bmp/BmpInfo4Bit.cpp
@@ -85,6 +85,6 @@ void BmpInfo4Bit::ByteArrayIndices( const int32_t X, const int32_t Y,
     VP_THROW( "y out of range" ); 
 #endif
 
-  ByteIndex = X/2 + (m_Height - 1 - Y)*m_RowLength;
+  ByteIndex = static_cast<uint32_t>(X/2) + static_cast<uint32_t>(m_Height - 1 - Y)*m_RowLength;
   BitIndex  = (X%2 == 0)? 0 : 4;
 }
